fix(repository): 修复多租户过滤器快照、UpdateAsync 死代码等仓储层问题
- 修复 ApplyTenantFilter 在 OnModelCreating 时将 TenantId 快照为常量的
  严重 bug，改为捕获 IEfTenantResolver 引用，每次查询动态读取，解决
  DbContextPool 场景下多租户隔离失效问题
- 修复 UpdateAsync(entity, expressions[]) 中 Detached 状态判断的死代码，
  现支持 detached 实体局部列更新
- 移除 EfTenantRepository.GetDbSet 中与全局过滤器重复的手动租户过滤
- 移除 EfTenantRepository.InsertAsync 中重复的 TenantId 手动赋值，
  统一由 TenantInterceptor 负责
- CountAsync / CountByTenantAsync / CountForCurrentTenantAsync 返回类型
  int 改为 long，GetAllAsync 新增 CancellationToken 参数（Breaking Change）
- DeletionAuditableFieldsInterceptor 级联软删除添加 visited 集合防循环引用
- 将 DisposeAction 内部类提取到独立文件；重命名 IAuditTimestamps.cs 为
  ITimestampAuditable.cs 以对齐接口名
- 新增 openspec 变更归档与能力规格文件

Made-with: Cursor

+++ src/Tenon.Repository/ITenantRepository.cs
@@ -72,7 +72,7 @@ public interface ITenantRepository<TEntity, in TKey, TTenantKey> : IRepository<T
     /// <param name="whereExpression">查询条件表达式</param>
     /// <param name="token">取消令牌</param>
     /// <returns>实体数量</returns>
-    Task<int> CountByTenantAsync(TTenantKey tenantId, Expression<Func<TEntity, bool>> whereExpression, CancellationToken token = default);
+    Task<long> CountByTenantAsync(TTenantKey tenantId, Expression<Func<TEntity, bool>> whereExpression, CancellationToken token = default);
     
     /// <summary>
     /// 异步获取当前租户下满足条件的实体数量
@@ -80,7 +80,7 @@ public interface ITenantRepository<TEntity, in TKey, TTenantKey> : IRepository<T
     /// <param name="whereExpression">查询条件表达式</param>
     /// <param name="token">取消令牌</param>
     /// <returns>实体数量</returns>
-    Task<int> CountForCurrentTenantAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken token = default);
+    Task<long> CountForCurrentTenantAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken token = default);
     
     /// <summary>
     /// 异步根据主键获取指定租户下的实体
